Move interleave_shuffle to reduce visibility
Untangle it. Call that conditioner.
